+++ demisto_sdk/commands/format/update_indicatorfields.py
@@ -2,6 +2,7 @@ from typing import Tuple
 
 from demisto_sdk.commands.common.hook_validations.incident_field import \
     IncidentFieldValidator
+from demisto_sdk.commands.common.tools import LOG_COLORS, print_color
 from demisto_sdk.commands.format.format_constants import (ERROR_RETURN_CODE,
                                                           SKIP_RETURN_CODE,
                                                           SUCCESS_RETURN_CODE)
@@ -13,7 +14,7 @@ class IndicatorFieldJSONFormat(BaseUpdateJSON):
 
         Attributes:
             input (str): the path to the file we are updating at the moment.
-            output (str): the desired file name to save the updated version of the YML to.
+            output (str): the desired file name to save the updated version of the JSON to.
     """
 
     def __init__(self, input: str = '', output: str = '', path: str = '', from_version: str = '', no_validate: bool = False):
@@ -21,15 +22,17 @@ class IndicatorFieldJSONFormat(BaseUpdateJSON):
 
     def run_format(self) -> int:
         try:
+            print_color(f'\n======= Updating file: {self.source_file} =======', LOG_COLORS.WHITE)
             super().update_json()
-            super().set_default_values_as_needed()
-            super().save_json_to_destination_file()
+            self.set_default_values_as_needed()
+            self.save_json_to_destination_file()
+
             return SUCCESS_RETURN_CODE
         except Exception:
             return ERROR_RETURN_CODE
 
     def format_file(self) -> Tuple[int, int]:
-        """Manager function for the integration YML updater."""
+        """Manager function for the indicator fields JSON updater."""
         format = self.run_format()
         if format:
             return format, SKIP_RETURN_CODE